backend/requirements.txt
# Core
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Database
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.0
authlib==1.3.0

# AI / LangChain ecosystem
langchain==0.2.5
langchain-google-genai==1.0.6
langchain-community==0.2.5
langgraph==0.1.14
langsmith==0.1.77
chromadb==0.5.3
tiktoken==0.7.0

# M-Pesa / HTTP
requests==2.32.3
aiohttp==3.9.5

# Data validation
pydantic==2.7.3
pydantic-settings==2.3.1
email-validator==2.1.1

# Utilities
python-dotenv==1.0.1
structlog==24.2.0
tenacity==8.3.0
celery==5.4.0
redis==5.0.4
flower==2.0.1

# Billing / Metering
stripe==9.12.0

# Testing
pytest==8.2.2
pytest-asyncio==0.23.7
httpx==0.27.0
